ISP20220437 * 調整ExchangeRate 資料來源, 資料型態 * import 不會 Query上一層InvNo資料 * 表身增加delete按鈕

+++ Database/dbo/Tables/BIRInvoice.sql
@@ -9,7 +9,7 @@
     [Approve]      VARCHAR (10)   NULL,
     [ApproveDate]  DATETIME       NULL,
     [InvDate]      DATE           NULL,
-    [ExchangeRate] SMALLINT       NULL,
+    [ExchangeRate] DECIMAL(18, 8)       NULL,
     [Remark]       NVARCHAR (500) NULL,
     CONSTRAINT [PK_BIRInvoice] PRIMARY KEY CLUSTERED ([ID] ASC)
 );
